+++ src/HolzShots/Input/Action/CaptureVideoCommand.cs
@@ -5,6 +5,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using HolzShots.Capture.Video;
+using HolzShots.Capture.Video.FFmpeg;
 using HolzShots.Composition.Command;
 
 namespace HolzShots.Input.Actions
